shif the format into the fmt module
